requirements.txt
streamlit==1.9.0
streamlit-aggrid==0.3.2
Pillow==9.1.1
